Use prototype methods, not property initializers
This appears to be a longstanding TypeScript feature request that's
not seeing much action: https://github.com/microsoft/TypeScript/issues/23911

What I wanted was to automatically gain the type inference for the
message type from the mapped type we're implementing, but what I
got was a lack of reuse of the functions themselves.

The extra redundancy is painful, but these really should be proper methods.