Revamp environment variable handling and the helper shell scripts
Finally made build.sh and restart.sh take their values from a local
(gitignored) .env file rather than hard-coding everything in. Updated
env_vars.ts to be capable of printing its output, with *opt-in* redaction
for secrets, to make it easier to verify that the running configuration
matches expectations.

Updated restart.sh to accept trailing arguments. When specified, will
run an interactive docker session using those arguments with a random
container name (for investigation. In particular, uWS segfaults if the
path to the certificates is invalid - and there are no docker logs on
a crashlooping container)