Add the ability to drain a lobby's rooms gracefully (#2)
* Add the ability to drain a lobby's rooms gracefully

Currently unconfigurable, but the behavior is as follows:
- lobby.drain() is called on SIGQUIT with a drop-dead time of 60 minutes
- any open rooms will receive a shutdown message immediately and every
  minute thereafter
- rooms that are actively in a run will be destroyed after 1 hour or
  5 minutes after the run ends, whichever comes first
- rooms that are NOT in a run will be destroyed after 5 minutes (if not
  destroyed sooner, e.g. by the users leaving)
- when all rooms are destroyed, the process will exit (disconnecting any remaining
  sockets in the process)
- in drain mode, new rooms cannot be created and new runs cannot be started

* Add env vars for tunable parameters. Warn when approaching payload limit.

* Revamp environment variable handling and the helper shell scripts

Finally made build.sh and restart.sh take their values from a local
(gitignored) .env file rather than hard-coding everything in. Updated
env_vars.ts to be capable of printing its output, with *opt-in* redaction
for secrets, to make it easier to verify that the running configuration
matches expectations.

Updated restart.sh to accept trailing arguments. When specified, will
run an interactive docker session using those arguments with a random
container name (for investigation. In particular, uWS segfaults if the
path to the certificates is invalid - and there are no docker logs on
a crashlooping container)

* Add a .env template file illustrating all the accepted vars

This codebase explicitly does *not* use the `dotenv` package,
since that package makes the operational behavior pretty opaque
and we expect to run within Docker. Instead, the env vars are
sourced into the `restart.sh` shell script and passed explicitly.

There is some minor logic to ease context switching between the
server environment and local development.

* Fix default values for all env vars

The new env file + shell script was passing empty strings for values,
which means that Node saw them as empty strings rather than undefined.
This caused the null-coalescing operator (`??`) to fail to fill the
values with defaults when the defaults were nonempty strings.

* Fix certificate mounting when using letsencrypt

* Fix JSDoc

* Add b/g switch scripts

* Implement build/restart; Clean up deployment-related scripts and data

Also make everything a bit more robust

* Bind local dev docker to 0.0.0.0 instead of 127.0.0.1

* Resolve room "drain" promise sooner than the timeout when possible